backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database & Cache
psycopg2-binary==2.9.7
redis==4.6.0
django-redis==5.4.0

# Authentication
djangorestframework-simplejwt==5.3.0
cryptography==41.0.7

# API Documentation
drf-spectacular==0.26.5

# Environment
python-decouple==3.8

# Utils
requests==2.31.0
python-dateutil==2.8.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3
